If the logger catalog exists, the log file in the log directory should also exist.  And vice versa: if the catalog does not yet exist, the log file also shouldn't exist.
We (currently) exit with a fatal error when this check fails.  If this
is too harsh (remember, the SQL and XQuery backends *must not start*
if the check fails), then the implementation of logger_fatal can be
changed.

This change stops users who think the log directory is optional (they
notice they just screwed up their database), and it helps in the
update from the Feb2009 release to the next release based on this
code.

Please, check the error messages to see whether they are clear enough.