Due to a change in the logdir location on the last feature release, we ended up in a position where an absolute location is not treated the same was as a relative location as logdir argument.  While the absolute location is treated in such a way that backwards compatability is kept, a relative location is treated in a new way to end up in some location relative to dbfarm/dbname.  In particular the relative location makes little sense to be changed (purely cosmetic), and is of limited freedom when moving up in the hierarchy (using ../).  Hence, the consensus has been reached that setting the value for the relative logdir is not necessary.  On the other hand, setting an absolute logdir is undesirable, as data and logdir go hand in hand, and without any of both the other is useless (hence putting the logs in volatile storage seems of no actual use).
For the next feature release, logdir can only be a relative path, and it
is hardwired to the current defaults (for backwards compatability).
Absolute paths are no longer allowed, requiring users to move their
logdirs to the relative location.  This means for SQL sql_logs, and for
XQuery xquery_logs.  A next feature release can then decide to change
these locations, given that it knows the previous ones.