It all started with a daily pop-up window on Windows reporting a failure in mclient.  This failure appeared to be an assertion failed which actually failed on all systems.  In turn, this assertion failed correctly since the server didn't adhere to the MAPI protocol in the particular case of the failing test.
The thing that went wrong in the server was a COPY INTO FROM STDIN
where a quote was missing in the data.  The server kept asking for
more data from the client, even though the client had already told it
that THERE IS NO MORE!

This checkin fixes the problems, and also problems with reporting
errors, which didn't always happen correctly.

Also, we now don't start more threads than there are columns.