Store string hash values in the string heap right before the string itself.  (Before that, the stridx_t, part of the double-elimination linked list, is stored.)
The hash is used when creating a hash table for a string bat,
hopefully improving the speed.

There is also code in there to store the string length before the hash
value, but this is not enabled in this version.  It would increase the
space requirements for probably very little (if any) gain.

This change is backward compatible with older versions of the
database, but not forward compatible.  The recent change to the GDK
version number also covers this change.

You will need to recompile everything after updating to this version
since the Heap structure was changed.