The following module implement a client interface to the Sphinx Fulltext Searchengine found at <http://sphinxsearch.com/> it requires the sphinxclient C api to be installed, and can be configured with --with-sphinxclient=...
When the server is started; --dbinit="include sphinx; ..." is required. In
order to use it in MonetDB5/SQL you would have to declare it as external
function:

create function sphinx_searchIndex(query STRING, idx STRING) RETURNS
        TABLE (id bigint) EXTERNAL NAME sphinx."sphinx_searchIndexWrap";


Typically Sphinx should gives a string lookup performance of around ~40ms,
where the searchd daemon MonetDB connects to at localhost returns a list
of documentids (32 or 64bit, depending on the Sphinx compiletime choice).

Currently the usage within SQL requires the user to review the generated
plan. It is possible that the external function will not get priority
over a join. You can trick the optimiser by:

select b, j from
	(select b from t where sphinx_searchIndex('bla', '*')) as x
		LEFT JOIN ...

Sphinx is able to index via an ODBC link. Please refer to the Sphinx
documentation to set up such index, and how to update it.