Removed mapi_check_query function.
It's been in the source for very long.  It seems to want to remove
comments from a given query (identified by #), but fails horribly in
doing so especially on multi-line queries.  This was revealed by latest
mapreduce code that uses comments to annotate blocks of statements for
debugging.

There seems to be absolutely no good reason at all to remove comments
from queries, the server should be able to handle them.  Besides from
unbreaking certain queries, by removing this function the speed of
mapi_query also increases as each query is scanned one time less now.