Put start scripts in autoload directory for modules installed.
For bug #2522, generate start scripts in the autoload directory of the
MonetDB5 libdir that are loaded from mal_init.  This way, only when a
module is available, it will be attempted to be loaded, avoiding
needless worrying warnings and allowing Testing to control which modules
are loaded and which not.