fixed ref-counting for BAT-of-BATs result created by TABLETcollect_bats
In CMDtablet_load() (tablet.load()) & CMDtablet_input() (tablet.input()),
the reference counting of the BAT-of-BATs result created by TABLETcollect_bats()
was not handled correctly. Instead of incrementing the logical refcount
and decrementing the physical refcount of the result BAT,
only BBPinref( x , TRUE ) was called on an uninitialized variable.
Fixed that.