More efficient internal TMsubcommit.
Internally, TMsubcommit is usually called between creating and filling
a new BAT and destroying that BAT.  That BAT is filled with names
which are looked up given a BAT ID, and then TMsubcommit converts
those names back to the selfsame BAT IDs.  This can be done more
efficiently.

We introduce a new function TMsubcommit_list with two parameters: an
array of BAT IDs, and a count.  Because of BBP-internal reasons, the
first element of the array is unused (and must be 0).  The count is
the number of BAT IDs, including the unused 0.

TMsubcommit now does the same as before, looking up BAT IDs and
putting them into an array.  Except now, TMsubcommit calls the new
function to do the "real" work.  This function can also be called from
other parts of the code.

By calling TMsubcommit_list, we now save on the creation and
destruction of temporary string BATs and concomitant locking, and we
save on the lookup of BAT names to find BAT IDs.

Note that this change only affects C code.  The MIL and MAL interfaces
remain identical.