Fix logic error.
If we close a result that is not the last one in the result sequence
received from the server, do not queue up an Xclose for it when all
tuples available in that result were already sent.  Due to what I
believe was a false ||, all results that were not the last were
unconditionally queued up for an Xclose command.  This causes problems
with the recent changes to debug listing (generating multiple result
sets, that the client erroneously wants to close) and the debugger
receiving Xclose commands which it doesn't expect.

I'm not sure anyone could ever encounter this bug, since if it would
have happened in normal use, it would just send an Xclose too much,
which the SQL server happily ignores.