Completed old change where BATs and reversed BATs can be named independently.
It used to be that when you named a BAT, the name applied to the BAT
as a whole, so when you then looked up the BAT by name you would
actually get its reverse (mirror).  That was changed long ago.  This
fix completes this change.  It could happen that unnamed mirrors ended
up in the BBP_hash table (used to look up BATs by name) which would
not be removed when the BAT was deleted, resulting in NULL pointer
dereferences.