revive investment in building hash table with equi-/point-selects
In Feb2010, changeset b8f00f4a09a5 removed the feature to build
a hash-table on-the-fly with equi-/point-selects, given that we
do not know for aure, whether the investment indeed pays off,
i.e., whether more equi-/point-selects on the same BAT will follow.

With this checkin, we revive that feature.
However, we avoid the investment in cases where benefits, i.e.,
future equi-/point-selects on the same BAT, are unlikely, i.e.,
on transient BATs.
We also avoid the investments in cases where the general benefit
is unclear, i.e., in case the atom type is narrow (1 or 2 bytes),
and hence the respective domain rather small, and in case the
hash-table does not fit in memory.