Introduce a detach argument for MT_create_thread.
It is necessary to distinguish between detached and joinable threads.
The former cannot be joined, whereas the latter actually have to be
joined in order to free resources.
By default, threads created by pthread are joinable.  In most places,
however, we do not join created threads, simply because we can't, or
never need to, until the end of the program.

Not joining joinable threads leaves behind a considerable amount of
memory, and registration in the threads pool.  As a result, thread
creation gets slower, and the mserver grows in memory.  Eventually, no
more threads can be created, as is the case in bug #2700.

We now specify per MT_create_thread if we want the thread to be joinable
or not, which in most cases is not, such that we don't leak, and can
continue handling queries, for which in our setup threads need to be
created.