Use MCgetClient to retrieve MAL client records
For a reason we do not yet completely understand, this used to work
fine, but now it doesn't.  Referencing mal_clients yields in another
address from modules, resulting in not the intended behaviour.  Use
MCgetClient to ensure that we retrieve the mal_client records from
mserver5, which does the main handling of clients.