use at least assert() to check whether malloc() & realloc() succeeded (when assertions are enabled)
mallocs can fail also for small sizes, and to prevent segfault, we should
always check whether a malloc succeeded or not, and properly handle the case
(error/exception) that is does not.
Here and for now, I just resorted to what's elswhere done in similar or
related code: use at least assert().
We should consider checking all malloc's properly, though ...