bootstrap: allow usage on Darwin and less-convenient platforms
Try to use glibtoolize instead of libtoolize when it exists, often the
case on Darwin/OSX machines.
Allow adding additional search paths to aclocal, for finding .m4 macro
files, which may not be in the aclocal default search path, as is
typically the case with a local install, or some separate less
convenient tool such as fink and macports.  Export M4DIRS as space
separated list of paths to use it.