move automatic modpath detection
It's better not to reinvent the proper modpath every time we need it,
but do it once instead.  At the same time, apparently find_include isn't
the only function that searches monet_mod_path, so by setting a value
for that variable early enough in the process we fix Mtest not being
able to list the available modules using the -s flag of mserver5.