RMTresolve: locate upstream merovingian via merovingian_uri
This is a fix for bug #2777.
The old approach of looking for mapi_socket in the dbfarm no longer
works, since sockets are stored in /tmp (by default) nowadays to make it
easier for clients to find them.  Extract the port from merovingian_uri
and pass that onto mapi_mapi (via mapi_resolve), which will connect to
merovingian using that port either through a UNIX or TCP socket.