memory detection: fix for Darwin
Use hw.memsize to obtain the physical memory installed in the system,
instead of hw.usermem which is unreliable above 2G and doesn't exactly
return what we do for other systems.

At the same time, also use sysctl on Darwin to obtain the pagesize from
the system (although the value happens to be the same as the fall-back
default).