PACMAN scheduler The scheduler has been distributed over the main scheduler and the workers. Whenever an instruction is finished by a worker, it chases the flow graph and execute all instructions that it feeds into exclusively. The main scheduler is notified, to ensure proper waiting for all results.
The effect of this algorithm is to ensure that each result produced
is immediately fed into another operator, thereby attempting to
avoid unnecessary IO of tmps.

Error handling in a worker blocks the main scheduler to feed new
instructions, but the worker has to continue until it reaches
a blocking instruction.