JDBC: rework PreparedStatement timestamp/time insertion
Timezones are important when inserting dates.  Because the server makes
a distinction between timezone-less and timezone-enabled time-fields,
the logic gets a bit complicated, since basically JDBC doesn't have to
do much when the server does all the timezone handling.  If not, JDBC
has to take into account the calendar (with timezone) given by the
caller, if any, as that calendar is supposed to reflect the timezone
state of the server being dealt with.