mclient: allow specifying time formatting
This makes one of my collegues happy, or mad.  Depends on how she looks
at it.
Implemented a simple overrride for the human-biased time formatting used
by mclient.  Instead of adjusting the time format to the value being
printed, allow forcing a certain representation (one of the three that
the human-biased method choses).